fix(db-explorer): validate table name against known tables, fix constant scope and portability
Agent-Logs-Url: https://github.com/lyquid/Romulus/sessions/6d3d98a6-b752-45ac-b2e8-0dc4ef63d0c2

+++ apps/gui/gui_app.cpp
@@ -902,7 +902,7 @@ void GuiApp::render_db_tab() {
   }
 
   ImGui::SameLine();
-  ImGui::TextDisabled("(%zu rows)", db_table_data_.rows.size());
+  ImGui::TextDisabled("(%lu rows)", static_cast<unsigned long>(db_table_data_.rows.size()));
 
   ImGui::Spacing();
